Validate python -> c # responses, ignore invalid best efforts

+++ src/backend/MyAIRunningMate.Python/app/domain/models/ingestion/activity_metrics.py
@@ -8,5 +8,5 @@ class ActivityMetrics(BaseModel):
     activity_metrics_ending_potential: Optional[int] = None
     activity_metrics_total_ascent: Optional[int] = None
     activity_metrics_total_descent: Optional[int] = None
-    activity_metrics_recovery_time: Optional[float] = None  
+    activity_metrics_recovery_time: Optional[int] = None  
     activity_metrics_num_laps: Optional[int] = None
