+++ src/backend/MyAIRunningMate.Python/app/domain/models/ingestion/time_series_record.py
@@ -9,7 +9,7 @@ class TimeSeriesRecord(BaseModel):
     tsr_heart_rate: Optional[int] = None
     tsr_cadence: Optional[int] = None
     tsr_distance_metres: Optional[float] = None
-    tsr_power: Optional[float] = None
+    tsr_power: Optional[int] = None
