+++ src/backend/MyAIRunningMate.Python/app/schemas/ingestion/best_effort_schema.py
@@ -5,7 +5,7 @@ from pydantic import BaseModel, ConfigDict
 
 
 class BestEffortSchema(BaseModel):
-    best_effort_distance_metres: Optional[float] = None
+    best_effort_distance_metres: Optional[int] = None
     best_effort_time_seconds: Optional[float] = None
     best_effort_is_personal_record: Optional[bool] = False
